refactor circular reference
