+++ api/handlers/datasources.py
@@ -1,10 +1,10 @@
 import httpx
 from database import datasource_repo
-from database.datasource_repo import create, datasource_exists, check_access
+from database.datasource_repo import create, datasource_exists
 from database.models import DataSource
 from fastapi import APIRouter, BackgroundTasks, Depends, HTTPException, Query
 from fastapi.responses import StreamingResponse
-from helpers.authorization import required_roles
+from helpers.authorization import required_roles, check_access
 from helpers.cipher import encrypt
 from helpers.urlmapping import ApiType, add_URL_sasToken
 from motor.core import AgnosticCollection
